Allow to download to input directory if specified
Backward behaviour is preserved if not input is provided,
it will use OS's tmp dir.

The motivation is to harden the action, to prevent injection,
if attacker is granted to write in /tmp,
the typical use case is to restrict to runner temporary dir.

This change has been tested using an other change in related downloader action.

Origin: https://github.com/ishworkh/container-image-artifact/pull/8
Relate-to: https://github.com/ishworkh/container-image-artifact-download/issues/7#issuecomment-2904751460
Signed-off-by: Philippe Coval <philippe.coval@silabs.com>